Deleted Home and changed to page.js. added carousel and altered styling being mobile friendly and shuffling photos to all the same size.
